+++ CHANGELOG.md
@@ -5,8 +5,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+## 7.1.1 - 2024-04-04
 ### Fixed
 - Updated CI library to latest
+- Updated dependencies. #INT-3274
+- Usage of RxJS deprecated operators. #INT-3274
+### Added
+- Support for Angular 15, 16 & 17. #INT-3274
+- Support for the OnPush change detection strategy. #INT-2974
+### Improved
+- Updated Storybook to v8, as well as now using CSFv3 components. #INT-3274
 
 ## 7.0.0 - 2022-06-27
 ### Added